# mdlsvc/gen/http/openapi.yaml
swagger: "2.0"
info:
    title: ""
    version: ""
host: localhost:80
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /api/packages:
        get:
            tags:
                - Packages
            summary: ListPackages Packages
            description: List the model packages in the given workspace
            operationId: Packages#ListPackages
            parameters:
                - name: work
                  in: query
                  description: Workspace identifier
                  required: true
                  type: string
                  minLength: 1
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/PackageResponse'
            schemes:
                - http
    /api/packages/files:
        get:
            tags:
                - Packages
            summary: ListPackageFiles Packages
            description: Get the DSL files and their content for the given model package
            operationId: Packages#ListPackageFiles
            parameters:
                - name: work
                  in: query
                  description: Workspace identifier
                  required: true
                  type: string
                  minLength: 1
                - name: dir
                  in: query
                  description: Path to directory containing a model package
                  required: true
                  type: string
                  minLength: 1
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/PackageFileResponse'
            schemes:
                - http
    /api/packages/layout:
        get:
            tags:
                - Packages
            summary: GetLayout Packages
            description: Streams the model layout JSON for the given package
            operationId: Packages#GetLayout
            parameters:
                - name: work
                  in: query
                  description: Workspace identifier
                  required: true
                  type: string
                  minLength: 1
                - name: dir
                  in: query
                  description: Path to directory containing a model package
                  required: true
                  type: string
                  minLength: 1
            responses:
                "200":
                    description: OK response.
            schemes:
                - http
    /api/packages/model:
        get:
            tags:
                - Packages
            summary: GetModelJSON Packages
            description: Streams the model JSON for the given package, see https://pkg.go.dev/goa.design/model/model#Model
            operationId: Packages#GetModelJSON
            parameters:
                - name: work
                  in: query
                  description: Workspace identifier
                  required: true
                  type: string
                  minLength: 1
                - name: dir
                  in: query
                  description: Path to directory containing a model package
                  required: true
                  type: string
                  minLength: 1
            responses:
                "200":
                    description: OK response.
            schemes:
                - http
    /api/packages/subscribe:
        get:
            tags:
                - Packages
            summary: Subscribe Packages
            description: Send model JSON on initial subscription and when the model package changes
            operationId: Packages#Subscribe
            parameters:
                - name: work
                  in: query
                  description: Workspace identifier
                  required: true
                  type: string
                  minLength: 1
                - name: dir
                  in: query
                  description: Path to directory containing a model package
                  required: true
                  type: string
                  minLength: 1
            responses:
                "101":
                    description: Switching Protocols response.
                    schema:
                        type: string
                        format: json
            schemes:
                - ws
...
definitions:
...
    FileLocatorResponse:
        title: FileLocatorResponse
        type: object
        properties:
            Dir:
                type: string
                description: Path to directory containing a model package
                example: src/repo/model
                minLength: 1
            Filename:
                type: string
                description: Name of DSL file
                example: model.go
                pattern: \.go$
            Workspace:
                type: string
                description: Workspace identifier
                example: my-workspace
                minLength: 1
        description: FileLocator is the location of a DSL file in a model package
        example:
            Dir: src/repo/model
            Filename: model.go
            Workspace: my-workspace
        required:
            - Filename
            - Workspace
            - Dir
    PackageFileResponse:
        title: PackageFileResponse
        type: object
        properties:
            Content:
                type: string
                description: DSL code
                example: |-
                    import . "goa.design/model/dsl"

                    var _ = Design(func() {})
                pattern: import . "goa.design/model/dsl"
                minLength: 58
            Locator:
                $ref: '#/definitions/FileLocatorResponse'
        example:
            Content: |-
                import . "goa.design/model/dsl"

                var _ = Design(func() {})
            Locator:
                Dir: src/repo/model
                Filename: model.go
                Workspace: my-workspace
        required:
            - Locator
            - Content
    PackageResponse:
        title: PackageResponse
        type: object
        properties:
            Dir:
                type: string
                description: Path to directory containing a model package
                example: src/repo/model
                minLength: 1
            ImportPath:
                type: string
                description: Design Go package import path
                example: goa.design/model/examples/basic/model
                pattern: ^([a-zA-Z0-9]+(-[a-zA-Z0-9]+)*\.)+[a-zA-Z]{2,}/[a-zA-Z0-9_\-]+/(/([a-zA-Z0-9_\-]+))*$
        example:
            Dir: src/repo/model
            ImportPath: goa.design/model/examples/basic/model
        required:
            - ImportPath
            - Dir
...