# mdlsvc/gen/http/packages/server/encode_decode.go
// Code generated by goa v3.13.2, DO NOT EDIT.
//
// Packages HTTP server encoders and decoders
//
// Command:
// $ goa gen goa.design/model/mdlsvc/design -o mdlsvc/

package server

import (
	"context"
	"net/http"
	"unicode/utf8"

	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
	types "goa.design/model/mdlsvc/gen/types"
)

// EncodeListPackagesResponse returns an encoder for responses returned by the
// Packages ListPackages endpoint.
func EncodeListPackagesResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.([]*types.Package)
		enc := encoder(ctx, w)
		body := NewListPackagesResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeListPackagesRequest returns a decoder for requests sent to the
// Packages ListPackages endpoint.
func DecodeListPackagesRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (any, error) {
	return func(r *http.Request) (any, error) {
		var (
			workspace string
			err       error
		)
		workspace = r.URL.Query().Get("work")
		if workspace == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("Workspace", "query string"))
		}
		if utf8.RuneCountInString(workspace) < 1 {
			err = goa.MergeErrors(err, goa.InvalidLengthError("Workspace", workspace, utf8.RuneCountInString(workspace), 1, true))
		}
		if err != nil {
			return nil, err
		}
		payload := NewListPackagesWorkspace(workspace)

		return payload, nil
	}
}

// EncodeListPackageFilesResponse returns an encoder for responses returned by
// the Packages ListPackageFiles endpoint.
func EncodeListPackageFilesResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.([]*types.PackageFile)
		enc := encoder(ctx, w)
		body := NewListPackageFilesResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeListPackageFilesRequest returns a decoder for requests sent to the
// Packages ListPackageFiles endpoint.
func DecodeListPackageFilesRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (any, error) {
	return func(r *http.Request) (any, error) {
		var (
			workspace string
			dir       string
			err       error
		)
		workspace = r.URL.Query().Get("work")
		if workspace == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("Workspace", "query string"))
		}
		if utf8.RuneCountInString(workspace) < 1 {
			err = goa.MergeErrors(err, goa.InvalidLengthError("Workspace", workspace, utf8.RuneCountInString(workspace), 1, true))
		}
		dir = r.URL.Query().Get("dir")
		if dir == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("Dir", "query string"))
		}
		if utf8.RuneCountInString(dir) < 1 {
			err = goa.MergeErrors(err, goa.InvalidLengthError("Dir", dir, utf8.RuneCountInString(dir), 1, true))
		}
		if err != nil {
			return nil, err
		}
		payload := NewListPackageFilesPackageLocator(workspace, dir)

		return payload, nil
	}
}

// DecodeSubscribeRequest returns a decoder for requests sent to the Packages
// Subscribe endpoint.
func DecodeSubscribeRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (any, error) {
	return func(r *http.Request) (any, error) {
		var (
			workspace string
			dir       string
			err       error
		)
		workspace = r.URL.Query().Get("work")
		if workspace == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("Workspace", "query string"))
		}
		if utf8.RuneCountInString(workspace) < 1 {
			err = goa.MergeErrors(err, goa.InvalidLengthError("Workspace", workspace, utf8.RuneCountInString(workspace), 1, true))
		}
		dir = r.URL.Query().Get("dir")
		if dir == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("Dir", "query string"))
		}
		if utf8.RuneCountInString(dir) < 1 {
			err = goa.MergeErrors(err, goa.InvalidLengthError("Dir", dir, utf8.RuneCountInString(dir), 1, true))
		}
		if err != nil {
			return nil, err
		}
		payload := NewSubscribePackageLocator(workspace, dir)

		return payload, nil
	}
}

// EncodeGetModelJSONResponse returns an encoder for responses returned by the
// Packages GetModelJSON endpoint.
func EncodeGetModelJSONResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		w.WriteHeader(http.StatusOK)
		return nil
	}
}

// DecodeGetModelJSONRequest returns a decoder for requests sent to the
// Packages GetModelJSON endpoint.
func DecodeGetModelJSONRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (any, error) {
	return func(r *http.Request) (any, error) {
		var (
			workspace string
			dir       string
			err       error
		)
		workspace = r.URL.Query().Get("work")
		if workspace == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("Workspace", "query string"))
		}
		if utf8.RuneCountInString(workspace) < 1 {
			err = goa.MergeErrors(err, goa.InvalidLengthError("Workspace", workspace, utf8.RuneCountInString(workspace), 1, true))
		}
		dir = r.URL.Query().Get("dir")
		if dir == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("Dir", "query string"))
		}
		if utf8.RuneCountInString(dir) < 1 {
			err = goa.MergeErrors(err, goa.InvalidLengthError("Dir", dir, utf8.RuneCountInString(dir), 1, true))
		}
		if err != nil {
			return nil, err
		}
		payload := NewGetModelJSONPackageLocator(workspace, dir)

		return payload, nil
	}
}

// EncodeGetLayoutResponse returns an encoder for responses returned by the
// Packages GetLayout endpoint.
func EncodeGetLayoutResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		w.WriteHeader(http.StatusOK)
		return nil
	}
}

// DecodeGetLayoutRequest returns a decoder for requests sent to the Packages
// GetLayout endpoint.
func DecodeGetLayoutRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (any, error) {
	return func(r *http.Request) (any, error) {
		var (
			workspace string
			dir       string
			err       error
		)
		workspace = r.URL.Query().Get("work")
		if workspace == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("Workspace", "query string"))
		}
		if utf8.RuneCountInString(workspace) < 1 {
			err = goa.MergeErrors(err, goa.InvalidLengthError("Workspace", workspace, utf8.RuneCountInString(workspace), 1, true))
		}
		dir = r.URL.Query().Get("dir")
		if dir == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("Dir", "query string"))
		}
		if utf8.RuneCountInString(dir) < 1 {
			err = goa.MergeErrors(err, goa.InvalidLengthError("Dir", dir, utf8.RuneCountInString(dir), 1, true))
		}
		if err != nil {
			return nil, err
		}
		payload := NewGetLayoutPackageLocator(workspace, dir)

		return payload, nil
	}
}

// marshalTypesPackageToPackageResponse builds a value of type *PackageResponse
// from a value of type *types.Package.
func marshalTypesPackageToPackageResponse(v *types.Package) *PackageResponse {
	res := &PackageResponse{
		ImportPath: v.ImportPath,
		Dir:        v.Dir,
	}

	return res
}

// marshalTypesPackageFileToPackageFileResponse builds a value of type
// *PackageFileResponse from a value of type *types.PackageFile.
func marshalTypesPackageFileToPackageFileResponse(v *types.PackageFile) *PackageFileResponse {
	res := &PackageFileResponse{
		Content: v.Content,
	}
	if v.Locator != nil {
		res.Locator = marshalTypesFileLocatorToFileLocatorResponse(v.Locator)
	}

	return res
}

// marshalTypesFileLocatorToFileLocatorResponse builds a value of type
// *FileLocatorResponse from a value of type *types.FileLocator.
func marshalTypesFileLocatorToFileLocatorResponse(v *types.FileLocator) *FileLocatorResponse {
	res := &FileLocatorResponse{
		Filename:  v.Filename,
		Workspace: v.Workspace,
		Dir:       v.Dir,
	}

	return res
}